Show views related to test/spec at the top of the list
With Test::Unit, view tests are found in the controller test file, so the best we can do is to show all views for the controller associated with the currently active controller spec at the top of the list.

RSpec, on the other hand, uses separate view specs, so here we can show the particular view associated with the currently active spec at the top of the list.